src/requirements.txt
python-dotenv
sqlalchemy
yfinance
pandas
psycopg2